+++ app/test/reading/create/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import CreateReadingTestPage from "@/features/test/reading/pages/create-reading-test";
+import CreateReadingTestPage from "@/features/test/pages/create-reading-test";
 import React from "react";
 
 const ReadingTest = () => {
