+++ features/test/pages/create-listening-test.tsx
@@ -8,7 +8,7 @@ import {
 } from "../form/create-listening-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { FormProvider, useForm } from "react-hook-form";
-import BaseForm from "../components/base-form";
+import BaseForm from "@/features/test/components/base-form";
 import { Separator } from "@/components/ui/separator";
 
 const CreateListeningTestPage = () => {
