+++ features/test/reading/components/questions/matching-information.tsx
@@ -14,7 +14,7 @@ import {
 import { Textarea } from "@/components/ui/textarea";
 import { Separator } from "@/components/ui/separator";
 import { Plus } from "lucide-react";
-import { indexToLetter } from "../../helpers/index-to-letter";
+import { indexToLetter } from "../../../../../helpers/index-to-letter";
 import QuestionBreakdown from "../question-breakdown";
 import { AnswerKeyField } from "../answer-key-field";
 import { PiCopyFill } from "react-icons/pi";
